changes on extraction

+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/spells/Extractor.java
@@ -25,6 +25,7 @@
 package com.shatteredpixel.shatteredpixeldungeon.items.spells;
 
 import com.shatteredpixel.shatteredpixeldungeon.Dungeon;
+import com.shatteredpixel.shatteredpixeldungeon.actors.buffs.Buff;
 import com.shatteredpixel.shatteredpixeldungeon.effects.Speck;
 import com.shatteredpixel.shatteredpixeldungeon.items.Item;
 import com.shatteredpixel.shatteredpixeldungeon.items.armor.Armor;
@@ -35,6 +36,7 @@ import com.shatteredpixel.shatteredpixeldungeon.items.scrolls.ScrollOfTransmutat
 import com.shatteredpixel.shatteredpixeldungeon.items.scrolls.ScrollOfUpgrade;
 import com.shatteredpixel.shatteredpixeldungeon.items.wands.Wand;
 import com.shatteredpixel.shatteredpixeldungeon.items.weapon.Weapon;
+import com.shatteredpixel.shatteredpixeldungeon.items.weapon.missiles.MissileWeapon;
 import com.shatteredpixel.shatteredpixeldungeon.messages.Messages;
 import com.shatteredpixel.shatteredpixeldungeon.sprites.ItemSpriteSheet;
 import com.shatteredpixel.shatteredpixeldungeon.utils.GLog;
@@ -56,11 +58,15 @@ public class Extractor extends InventorySpell {
 	@Override
 	protected void onItemSelected(Item item) {
 
-		Dungeon.level.drop(new ScrollOfUpgrade().quantity(item.level()), curUser.pos).sprite.drop();
-		item.level(0);
+		if (!(item instanceof MissileWeapon)) {
+			Dungeon.level.drop(new ScrollOfUpgrade().quantity(item.level()), curUser.pos).sprite.drop();
+			item.level(0);
 
-		GLog.p(Messages.get(this, "extracted"));
-		curUser.sprite.emitter().start(Speck.factory(Speck.UP), 0.2f, 10);
+			GLog.p(Messages.get(this, "extracted"));
+			curUser.sprite.emitter().start(Speck.factory(Speck.UP), 0.2f, 10);
+		} else {
+
+		}
 	}
 	
 	@Override
